refactor(appmenu): changed menues file structure
