+++ src/components/appMenu/CircularMenu/CircularMenu.module.scss
