+++ src/components/appMenu/CircularMenu/CircularMenuItem.module.scss
@@ -1,4 +1,4 @@
-@import '../../style/mixins.scss';
+@import '../../../style/mixins.scss';
 
 .menu_item {
   box-shadow: 0px 0px 13px #ffffff54;
