+++ src/components/appMenu/MobileMenu/MobileMenu.module.scss
@@ -1,4 +1,4 @@
-@import '../../style/mixins.scss';
+@import '../../../style/mixins.scss';
 
 .mobileMenu {
   position: fixed;
