+++ src/layouts/Main.tsx
@@ -12,9 +12,9 @@ import TimeFooter from 'src/features/TimeFooter/TimeFooter';
 import { Networks } from 'src/types/networks';
 import { CHAIN_ID } from 'src/constants/config';
 import { Link } from 'react-router-dom';
-import CircularMenu from 'src/components/appMenu/CircularMenu';
+import CircularMenu from 'src/components/appMenu/CircularMenu/CircularMenu';
 import TimeHistory from 'src/features/TimeHistory/TimeHistory';
-import MobileMenu from 'src/components/appMenu/MobileMenu';
+import MobileMenu from 'src/components/appMenu/MobileMenu/MobileMenu';
 import graphDataPrepared from '../pages/oracle/landing/graphDataPrepared.json';
 import stylesOracle from '../pages/oracle/landing/OracleLanding.module.scss';
 import SenseButton from '../features/sense/ui/SenseButton/SenseButton';
